back/requirements.txt
pdfminer.six
requests
flask
flask-cors
